+++ src/calkit/__init__.py
@@ -62,4 +62,4 @@ __all__ = [
     "write_reminder_shortcut",
 ]
 
-__version__ = "0.4.0"
+__version__ = "0.4.1"